creek-backend/requirements.txt
# Core
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=1.10.7,<2.0.0

# Audio Processing
sounddevice>=0.4.6
soundfile>=0.12.1
webrtcvad>=2.0.10
librosa>=0.10.0

# Machine Learning
torch>=2.0.0
torchaudio>=2.0.0
espnet-tts-pytorch>=0.1.0
huggingface-hub>=0.13.0
transformers>=4.28.0

# Document Processing
python-docx>=0.8.11
python-pptx>=0.6.21

# Utilities
numpy>=1.23.5
python-magic>=0.4.27
tqdm>=4.65.0
python-dateutil>=2.8.2
requests>=2.28.2
httpx>=0.24.0
